feat(about-me): add About Me screen with Mac OS-like UI components
Added About Me screen containing:
- Accordion with titles and answers
- Profile photo and name
- Map
- Dynamic years of experience counter
