--- src/app/hooks/useHorizonatScroll.ts
@@ -1,19 +1,0 @@
-import { useEffect } from "react";
-
-export function useHorizontalScroll(ref: React.RefObject<HTMLDivElement | null>) {
-  useEffect(() => {
-    const container = ref.current;
-    if (!container) return;
-
-    const handleWheel = (e: WheelEvent) => {
-      if (e.deltaY !== 0) {
-        e.preventDefault();
-        const speed = 0.5;
-        container.scrollLeft += e.deltaY * speed;
-      }
-    };
-
-    container.addEventListener("wheel", handleWheel, { passive: false });
-    return () => container.removeEventListener("wheel", handleWheel);
-  }, [ref]);
-}
