+++ src/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
-import Sidebar from './components/molecules/shared/Sidebar';
 import './globals.css';
-import Background from './components/molecules/shared/Background';
+import Sidebar from '@/components/molecules/shared/Sidebar';
+import Background from '@/components/molecules/shared/Background';
+
 
 const geistSans = Geist({
     variable: '--font-geist-sans',
@@ -29,7 +30,7 @@ export default function RootLayout({
             <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-gray-950`}>
                 <div className="flex min-h-screen">
                     <Sidebar />
-                    <Background />
+                    
                     <main className="flex-1">{children}</main>
                 </div>
             </body>
